pyproject.toml
[project]
name = "ai-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "bs4>=0.0.2",
    "email-validator>=2.0",
    "fastapi>=0.136.1",
    "itsdangerous>=2.2.0",
    "langdetect>=1.0.9",
    "openai>=2.33.0",
    "playwright>=1.58.0",
    "psycopg[binary]>=3.3.4",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.8",
    "requests>=2.33.1",
    "sqladmin>=0.25.0",
    "sqlalchemy[asyncio]>=2.0",
    "tiktoken>=0.12.0",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.46.0",
]
